Remove the Images dependencies
We would like to use this setup for spatial data .
Therefore, we don't need the setup for images.
I also renamed the package and generated a new UUID.
