+++ Project.toml
@@ -1,36 +1,17 @@
-name = "TestImages"
-uuid = "5e47fb64-e119-507b-a336-dd2b206d9990"
-version = "1.7.0"
+name = "SpatialTestData"
+uuid = "d9246668-2619-4ebb-87e5-973a747205de"
+version = "0.0.1"
 
 [deps]
-AxisArrays = "39de3d68-74b9-583c-8d2d-e117c070f3a9"
-ColorTypes = "3da002f7-5984-5a60-b8a6-cbb66c0b333f"
-FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
-ImageIO = "82e4d734-157c-48bb-816b-45c225c6df19"
-ImageMagick = "6218d12a-5da1-5696-b52f-db25d2ecc6d1"
-OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 StringDistances = "88034a9c-02f8-509d-84a9-84ec65e18404"
 
 [compat]
-AxisArrays = "0.3, 0.4"
-ColorTypes = "0.8, 0.9, 0.10, 0.11"
-ImageIO = "0.3, 0.4, 0.5, 0.6"
-ImageMagick = "1"
-FileIO = "1"
-OffsetArrays = "1"
 StringDistances = "0.10, 0.11"
 julia = "1.3"
 
 [extras]
-Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
-FixedPointNumbers = "53c48c17-4a7d-5ca2-90c5-79b7896eea93"
-ImageContrastAdjustment = "f332f351-ec65-5f6a-b3d1-319c6670881a"
-ImageMagick = "6218d12a-5da1-5696-b52f-db25d2ecc6d1"
-OMETIFF = "2d0ec36b-e807-5756-994b-45af29551fcf"
-ReferenceTests = "324d217c-45ce-50fc-942e-d289b448e8cf"
-Suppressor = "fd094767-a336-5f1f-9728-57cf17d0bbfb"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Colors", "FixedPointNumbers", "ImageMagick", "ImageContrastAdjustment", "OMETIFF", "ReferenceTests", "Suppressor", "Test"]
+test = ["Test"]
